Restructure libs/ into logical subpackages
Organized libs/ into subdirectories for better maintainability:

libs/
├── core/       - Shape, commands (undo/redo), settings
├── formats/    - Pascal VOC, YOLO, CreateML I/O, LabelFile
├── widgets/    - Canvas, gallery, toolbar, dialogs, etc.
├── utils/      - Constants, stringBundle, styles, helpers
└── resources.py (stays at root - auto-generated)

Updated all imports throughout:
- labelImgPlusPlus.py uses new paths
- Internal libs imports updated
- Test imports updated

Also updated Makefile test command for proper discovery.
