+++ libs/utils/hashableQListWidgetItem.py
