+++ libs/widgets/__init__.py
@@ -1,0 +1,19 @@
+# libs/widgets/__init__.py
+"""UI widgets and components."""
+
+from libs.widgets.canvas import Canvas
+from libs.widgets.galleryWidget import GalleryWidget
+from libs.widgets.toolBar import ToolBar
+from libs.widgets.statsWidget import StatsWidget
+from libs.widgets.labelDialog import LabelDialog
+from libs.widgets.colorDialog import ColorDialog
+from libs.widgets.zoomWidget import ZoomWidget
+from libs.widgets.lightWidget import LightWidget
+from libs.widgets.combobox import ComboBox
+from libs.widgets.default_label_combobox import DefaultLabelComboBox
+
+__all__ = [
+    'Canvas', 'GalleryWidget', 'ToolBar', 'StatsWidget',
+    'LabelDialog', 'ColorDialog', 'ZoomWidget', 'LightWidget',
+    'ComboBox', 'DefaultLabelComboBox',
+]
