+++ libs/widgets/lightWidget.py
