+++ libs/widgets/zoomWidget.py
