+++ tests/core/test_settings.py
@@ -6,9 +6,8 @@ import tempfile
 import unittest
 
 dir_name = os.path.abspath(os.path.dirname(__file__))
-libs_path = os.path.join(dir_name, '..', '..', 'libs')
-sys.path.insert(0, libs_path)
-from settings import Settings
+sys.path.insert(0, os.path.join(dir_name, '..', '..'))
+from libs.core.settings import Settings
 
 
 class TestSettings(unittest.TestCase):
