+++ tests/core/test_shape.py
@@ -14,7 +14,7 @@ sys.path.insert(0, os.path.join(dir_name, '..', '..', 'libs'))
 from PyQt5.QtCore import QPointF
 from PyQt5.QtWidgets import QApplication
 
-from libs.shape import Shape
+from libs.core.shape import Shape
 
 # Create QApplication for tests
 app = QApplication.instance() or QApplication(sys.argv)
