+++ tests/widgets/test_label_dialog.py
@@ -12,7 +12,7 @@ dir_name = os.path.abspath(os.path.dirname(__file__))
 sys.path.insert(0, os.path.join(dir_name, '..', '..'))
 
 from PyQt5.QtWidgets import QApplication
-from libs.labelDialog import LabelDialog
+from libs.widgets.labelDialog import LabelDialog
 
 
 class TestLabelDialogFilter(unittest.TestCase):
